// src/generated/rustplus.ts
export interface AppRequest {
  seq: number;
  playerId: string;
  playerToken: number;
  entityId?: number | undefined;
  getInfo?: AppEmpty | undefined;
  getTime?: AppEmpty | undefined;
  getMap?: AppEmpty | undefined;
  getTeamInfo?: AppEmpty | undefined;
  getTeamChat?: AppEmpty | undefined;
  sendTeamMessage?: AppSendMessage | undefined;
  getEntityInfo?: AppEmpty | undefined;
  setEntityValue?: AppSetEntityValue | undefined;
  checkSubscription?: AppEmpty | undefined;
  setSubscription?: AppFlag | undefined;
  getMapMarkers?: AppEmpty | undefined;
  promoteToLeader?: AppPromoteToLeader | undefined;
  getClanInfo?: AppEmpty | undefined;
  setClanMotd?: AppSendMessage | undefined;
  getClanChat?: AppEmpty | undefined;
  sendClanMessage?: AppSendMessage | undefined;
  getNexusAuth?: AppGetNexusAuth | undefined;
  cameraSubscribe?: AppCameraSubscribe | undefined;
  cameraUnsubscribe?: AppEmpty | undefined;
  cameraInput?: AppCameraInput | undefined;
}

export interface AppMessage {
  response?: AppResponse | undefined;
  broadcast?: AppBroadcast | undefined;
}

export interface AppResponse {
  seq: number;
  success?: AppSuccess | undefined;
  error?: AppError | undefined;
  info?: AppInfo | undefined;
  time?: AppTime | undefined;
  map?: AppMap | undefined;
  teamInfo?: AppTeamInfo | undefined;
  teamChat?: AppTeamChat | undefined;
  entityInfo?: AppEntityInfo | undefined;
  flag?: AppFlag | undefined;
  mapMarkers?: AppMapMarkers | undefined;
  clanInfo?: AppClanInfo | undefined;
  clanChat?: AppClanChat | undefined;
  nexusAuth?: AppNexusAuth | undefined;
  cameraSubscribeInfo?: AppCameraInfo | undefined;
}

export interface AppBroadcast {
  teamChanged?: AppTeamChanged | undefined;
  teamMessage?: AppNewTeamMessage | undefined;
  entityChanged?: AppEntityChanged | undefined;
  clanChanged?: AppClanChanged | undefined;
  clanMessage?: AppNewClanMessage | undefined;
  cameraRays?: AppCameraRays | undefined;
}

export interface AppEmpty {
}

export interface AppSendMessage {
  message: string;
}

export interface AppSetEntityValue {
  value: boolean;
}

export interface AppPromoteToLeader {
  steamId: string;
}

export interface AppGetNexusAuth {
  appKey: string;
}

export interface AppSuccess {
}

export interface AppError {
  error: string;
}

export interface AppFlag {
  value: boolean;
}

export interface AppInfo {
  name: string;
  headerImage: string;
  url: string;
  map: string;
  mapSize: number;
  wipeTime: number;
  players: number;
  maxPlayers: number;
  queuedPlayers: number;
  seed?: number | undefined;
  salt?: number | undefined;
  logoImage?: string | undefined;
  nexus?: string | undefined;
  nexusId?: number | undefined;
  nexusZone?: string | undefined;
}

export interface AppTime {
  dayLengthMinutes: number;
  timeScale: number;
  sunrise: number;
  sunset: number;
  time: number;
}

export interface AppMap {
  width: number;
  height: number;
  jpgImage: Uint8Array;
  oceanMargin: number;
  monuments: AppMap_Monument[];
  background?: string | undefined;
}

export interface AppMap_Monument {
  token: string;
  x: number;
  y: number;
}

export interface AppEntityInfo {
  type: AppEntityType;
  payload?: AppEntityPayload | undefined;
}

export interface AppEntityPayload {
  value?: boolean | undefined;
  items: AppEntityPayload_Item[];
  capacity?: number | undefined;
  hasProtection?: boolean | undefined;
  protectionExpiry?: number | undefined;
}

export interface AppEntityPayload_Item {
  itemId: number;
  quantity: number;
  itemIsBlueprint: boolean;
}

export interface AppTeamInfo {
  leaderSteamId: string;
  members: AppTeamInfo_Member[];
  mapNotes: AppTeamInfo_Note[];
  leaderMapNotes: AppTeamInfo_Note[];
}
// ...
export interface AppMarker {
  id: number;
  type: AppMarkerType;
  x: number;
  y: number;
  steamId?: string | undefined;
  rotation?: number | undefined;
  radius?: number | undefined;
  color1?: Vector4 | undefined;
  color2?: Vector4 | undefined;
  alpha?: number | undefined;
  name?: string | undefined;
  outOfStock?: boolean | undefined;
  sellOrders: AppMarker_SellOrder[];
}
// ...
export interface AppNexusAuth {
  serverId: string;
  playerToken: number;
}
// ...
export interface AppCameraInfo {
  width: number;
  height: number;
  nearPlane: number;
  farPlane: number;
  controlFlags: number;
}
// ...
export const AppMessage: MessageFns<AppMessage> = {
  encode(message: AppMessage, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.response !== undefined) {
      AppResponse.encode(message.response, writer.uint32(10).fork()).join();
    }
    if (message.broadcast !== undefined) {
      AppBroadcast.encode(message.broadcast, writer.uint32(18).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AppMessage {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAppMessage();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.response = AppResponse.decode(reader, reader.uint32());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.broadcast = AppBroadcast.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<AppMessage>, I>>(base?: I): AppMessage {
    return AppMessage.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<AppMessage>, I>>(object: I): AppMessage {
    const message = createBaseAppMessage();
    message.response = (object.response !== undefined && object.response !== null)
      ? AppResponse.fromPartial(object.response)
      : undefined;
    message.broadcast = (object.broadcast !== undefined && object.broadcast !== null)
      ? AppBroadcast.fromPartial(object.broadcast)
      : undefined;
    return message;
  },
};
// ...
export const AppMap: MessageFns<AppMap> = {
  encode(message: AppMap, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.width !== 0) {
      writer.uint32(8).uint32(message.width);
    }
    if (message.height !== 0) {
      writer.uint32(16).uint32(message.height);
    }
    if (message.jpgImage.length !== 0) {
      writer.uint32(26).bytes(message.jpgImage);
    }
    if (message.oceanMargin !== 0) {
      writer.uint32(32).int32(message.oceanMargin);
    }
    for (const v of message.monuments) {
      AppMap_Monument.encode(v!, writer.uint32(42).fork()).join();
    }
    if (message.background !== undefined && message.background !== "") {
      writer.uint32(50).string(message.background);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AppMap {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAppMap();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.width = reader.uint32();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.height = reader.uint32();
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.jpgImage = reader.bytes();
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.oceanMargin = reader.int32();
          continue;
        }
        case 5: {
          if (tag !== 42) {
            break;
          }

          message.monuments.push(AppMap_Monument.decode(reader, reader.uint32()));
          continue;
        }
        case 6: {
          if (tag !== 50) {
            break;
          }

          message.background = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<AppMap>, I>>(base?: I): AppMap {
    return AppMap.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<AppMap>, I>>(object: I): AppMap {
    const message = createBaseAppMap();
    message.width = object.width ?? 0;
    message.height = object.height ?? 0;
    message.jpgImage = object.jpgImage ?? new Uint8Array(0);
    message.oceanMargin = object.oceanMargin ?? 0;
    message.monuments = object.monuments?.map((e) => AppMap_Monument.fromPartial(e)) || [];
    message.background = object.background ?? "";
    return message;
  },
};
// ...
export const AppEntityInfo: MessageFns<AppEntityInfo> = {
  encode(message: AppEntityInfo, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.type !== 1) {
      writer.uint32(8).int32(message.type);
    }
    if (message.payload !== undefined) {
      AppEntityPayload.encode(message.payload, writer.uint32(26).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AppEntityInfo {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAppEntityInfo();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.type = reader.int32() as any;
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.payload = AppEntityPayload.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<AppEntityInfo>, I>>(base?: I): AppEntityInfo {
    return AppEntityInfo.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<AppEntityInfo>, I>>(object: I): AppEntityInfo {
    const message = createBaseAppEntityInfo();
    message.type = object.type ?? 1;
    message.payload = (object.payload !== undefined && object.payload !== null)
      ? AppEntityPayload.fromPartial(object.payload)
      : undefined;
    return message;
  },
};
// ...
export const AppCameraInfo: MessageFns<AppCameraInfo> = {
  encode(message: AppCameraInfo, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.width !== 0) {
      writer.uint32(8).int32(message.width);
    }
    if (message.height !== 0) {
      writer.uint32(16).int32(message.height);
    }
    if (message.nearPlane !== 0) {
      writer.uint32(29).float(message.nearPlane);
    }
    if (message.farPlane !== 0) {
      writer.uint32(37).float(message.farPlane);
    }
    if (message.controlFlags !== 0) {
      writer.uint32(40).int32(message.controlFlags);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AppCameraInfo {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAppCameraInfo();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.width = reader.int32();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.height = reader.int32();
          continue;
        }
        case 3: {
          if (tag !== 29) {
            break;
          }

          message.nearPlane = reader.float();
          continue;
        }
        case 4: {
          if (tag !== 37) {
            break;
          }

          message.farPlane = reader.float();
          continue;
        }
        case 5: {
          if (tag !== 40) {
            break;
          }

          message.controlFlags = reader.int32();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<AppCameraInfo>, I>>(base?: I): AppCameraInfo {
    return AppCameraInfo.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<AppCameraInfo>, I>>(object: I): AppCameraInfo {
    const message = createBaseAppCameraInfo();
    message.width = object.width ?? 0;
    message.height = object.height ?? 0;
    message.nearPlane = object.nearPlane ?? 0;
    message.farPlane = object.farPlane ?? 0;
    message.controlFlags = object.controlFlags ?? 0;
    return message;
  },
};